Add minmax limits to column size
Uses the minmax function to replace the 1fr in the repeat function with a
column size that has the minimum width of 90px and the maximum width of 1fr,
Resize the preview panel to see the effect. Shrinking panel is limited
to 90 px collumns. Expanding panel increase to all equal 1fr size columns.