Replace grids 3 fixed columns with auto-fill
Under container class, add auto-fill with repeat to fill the grid
with columns that have a minimum width of 60px and max of 1fr.
This allows framework to stretch by adding more columns as needed
to suite the cell columns size restriction.  This example compares
the grid on top with auto-fill to the grid on bottom with fixed
number of columns.