Create a function to modify JSON object
Creates a function which takes an album's id (like 2548), a property prop
(like "artist" or "tracks"), and a value (like "Addicted to Love") to modify
the data in this collection.

If prop isn't "tracks" and value isn't empty (""), updates or set the value
for that record album's property.

function must always return the entire collection object.