Modify lookupProfile function
LookUpProfile function takes name and a property (prop) as arguments

The function should checks if name is an actual contact's firstName and
the given property (prop) is a property of that contact.

If both are true, then return the "value" of that property.

If name does not correspond to any contacts then return "No such contact"

If prop does not correspond to any valid properties of a contact found
to match name then return "No such property"